+++ PluginManager/Interfaces/Files/IWritableFileSystemInfo.cs
@@ -1,8 +1,12 @@
 namespace Nanoray.PluginManager;
 
-public interface IWritableFileSystemInfo<TFileInfo, TDirectoryInfo> : IFileSystemInfo<TFileInfo, TDirectoryInfo>
+public interface IWritableFileSystemInfo : IFileSystemInfo
+{
+	void Delete();
+}
+
+public interface IWritableFileSystemInfo<TFileInfo, TDirectoryInfo> : IFileSystemInfo<TFileInfo, TDirectoryInfo>, IWritableFileSystemInfo
 	where TFileInfo : class, IWritableFileInfo<TFileInfo, TDirectoryInfo>
 	where TDirectoryInfo : class, IWritableDirectoryInfo<TFileInfo, TDirectoryInfo>
 {
-	void Delete();
 }